+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed timestamp in GraphQL API
+- Fixed BigInt in GraphQL API
 
 ## [1.0.0] 2023-08-22